packages/dffmpeg-coordinator/pyproject.toml
[project]
name = "dffmpeg-coordinator"
version = "0.1.0"
authors = [
    { name="Bryan Wyatt", email="brwyatt@gmail.com" },
]
description = "Coordinator application components for dffmpeg"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aiosqlite",
    "dffmpeg-common",
    "fastapi",
    "pydantic",
    "python-ulid[pydantic]",
    "pyyaml",
    "uvicorn",
    "sqlalchemy",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dffmpeg.*"]
namespaces = true

[project.optional-dependencies]
test = [
    "pytest",
    "black",
    "isort",
    "flake8",
]
dev = [
    "dffmpeg-coordinator[test]",
    "httpx",
    "ipython",
]

[tool.setuptools.package-data]
"*" = ["*.sql"]

[project.scripts]

[project.entry-points."dffmpeg.db.auth"]
sqlite = "dffmpeg.coordinator.db.auth.sqlite:SQLiteAuthRepository"

[project.entry-points."dffmpeg.db.jobs"]
sqlite = "dffmpeg.coordinator.db.jobs.sqlite:SQLiteJobRepository"

[project.entry-points."dffmpeg.db.messages"]
sqlite = "dffmpeg.coordinator.db.messages.sqlite:SQLiteMessageRepository"

[project.entry-points."dffmpeg.db.workers"]
sqlite = "dffmpeg.coordinator.db.workers.sqlite:SQLiteWorkerRepository"

[project.entry-points."dffmpeg.transports.server"]
http_polling = "dffmpeg.coordinator.transports.http_polling:HTTPPollingTransport"
